Moved handleStartTraining under ./actions
